requirements.txt
fastapi[standard]
langchain
langchain-openai
langchain-qdrant
langgraph

# [DB Client]
qdrant-client
python-dotenv
uvicorn
pytest
httpx
pydantic

# 의존성 임시 추가
langchain-google-genai

sentence-transformers
